Graph styles pane architecture refactor
