+++ src/framework/graph-styles-pane/graph-styles-pane.component.scss
