+++ src/module/schema/schema-page.component.html
@@ -42,7 +42,7 @@
                 }
               }
             </div>
-            <ts-graph-customisation-panel resizable [percent]="panelSizes[3]" (percentChange)="onPanelResize(3, $event)" [visualiser]="state.visualiser.visualiser"/>
+            <ts-graph-styles-pane resizable [percent]="panelSizes[3]" (percentChange)="onPanelResize(3, $event)" [visualiser]="state.visualiser.visualiser"/>
           </div>
         </div>
       }
